Refactor OmgUtil package
