+++ src/common/hooks/useRootchainTracker.js
@@ -1,7 +1,7 @@
-import { useEffect, useCallback, useState } from 'react'
-import { TransactionActionTypes } from 'common/constants'
 import Config from 'react-native-config'
-import { NotificationMessages } from 'common/constants'
+import { useEffect, useCallback, useState } from 'react'
+import { NotificationMessages, TransactionActionTypes } from 'common/constants'
+import { OmgUtil } from 'common/utils'
 import { Plasma } from 'common/blockchain'
 
 const getConfirmationsThreshold = tx => {
@@ -75,7 +75,7 @@ const useRootchainTracker = wallet => {
 
   const track = useCallback(async () => {
     const latestPendingTx = pendingRootchainTxs.slice(-1).pop()
-    const receipt = await Plasma.waitForRootchainTransaction({
+    const receipt = await OmgUtil.waitForRootchainTransaction({
       hash: latestPendingTx.hash,
       intervalMs: 3000,
       confirmationThreshold: getConfirmationsThreshold(latestPendingTx),
