+++ src/common/utils/index.js
@@ -1,3 +1,4 @@
+import OmgUtil from '@omisego/react-native-omg-js'
 import * as Price from './price'
 import * as Storage from './storage'
 import * as Datetime from './datetime'
@@ -30,6 +31,7 @@ export {
   GasEstimator,
   Mapper,
   Notification,
+  OmgUtil,
   Identicon,
   Parser,
   Polling,
